bpf: move EGW-related structs to egress_gateway.h
Untangle common.h some more.
